Reorganize Android codebase structure
