--- client/app/src/main/java/group01/smartcar/client/Status.java
@@ -1,7 +1,0 @@
-package group01.smartcar.client;
-
-public enum Status {
-    ACTIVE,
-    INACTIVE,
-    PAUSED
-}
