+++ client/app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:rotation="0"
-    tools:context=".MainActivity">
+    tools:context=".activity.LoginActivity">
 
     <VideoView
         android:id="@+id/videoView"
